requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
yfinance
jupyterlab
ipykernel
black
ruff
pytest
